Split chart functionality into separate module
