+++ index.html
@@ -83,6 +83,7 @@
     <script src="https://cdn.jsdelivr.net/npm/chart.js@2.8.0"></script>
     <script src="https://cdn.jsdelivr.net/npm/chartjs-plugin-datalabels@0.5.0"></script>
     <script src="tokenData.js" type="module"></script>
+    <script src="tokenCharts.js" type="module"></script>
     <script src="token.js" type="module"></script>
   </body>
 </html>
